Update README.md
Java based simple E-commerce website. Which perform the simply create and search operation. In this project mainly two user as buyer and seller. Where buyer user bye the product and seller user add the product into the database. For this project development we use Java and Java Fxml for Gui and MySql database is used for stored the data.